chore[Op#54459]: rename `Projects::Scopes::AvailableCustomFields` to include negation scope
